backend/requirements.txt
bcrypt==4.2.1
PyJWT==2.10.1
cryptography==44.0.2

